build: add build tag to disable eBPF
we are using runc to read the cgroup configuration, if eBPF is enabled
it pulls in the cilium/ebpf dependency that increases the binary size
by ~1 MB.
