+++ README.md
@@ -65,9 +65,10 @@ e.g. to disable seccomp:
 make BUILDTAGS=""
 ```
 
-| Build Tag | Feature                            | Enabled by default | Dependency |
-|-----------|------------------------------------|--------------------|------------|
-| seccomp   | Syscall filtering                  | yes                | libseccomp |
+| Build Tag      | Feature                            | Enabled by default | Dependency |
+|----------------|------------------------------------|--------------------|------------|
+| seccomp        | Syscall filtering                  | yes                | libseccomp |
+| runc_no_ebpf   | Disable eBPF support               | no                 |            |
 
 The following build tags were used earlier, but are now obsoleted:
  - **nokmem** (since runc v1.0.0-rc94 kernel memory settings are ignored)
