+++ libcontainer/cgroups/devices/ebpf_linux.go
@@ -1,3 +1,5 @@
+//go:build !runc_no_ebpf
+
 package devices
 
 import (
@@ -7,6 +9,8 @@ import (
 	"runtime"
 	"sync"
 	"unsafe"
+
+	"github.com/opencontainers/runc/libcontainer/configs"
 
 	"github.com/cilium/ebpf"
 	"github.com/cilium/ebpf/asm"
@@ -250,4 +254,25 @@ func loadAttachCgroupDeviceFilter(insts asm.Instructions, license string, dirFd 
 		}
 	}
 	return closer, nil
+}
+
+func setV2(dirPath string, r *configs.Resources) error {
+	if r.SkipDevices {
+		return nil
+	}
+	insts, license, err := deviceFilter(r.Devices)
+	if err != nil {
+		return err
+	}
+	dirFD, err := unix.Open(dirPath, unix.O_DIRECTORY|unix.O_RDONLY, 0o600)
+	if err != nil {
+		return fmt.Errorf("cannot get dir FD for %s", dirPath)
+	}
+	defer unix.Close(dirFD)
+	if _, err := loadAttachCgroupDeviceFilter(insts, license, dirFD); err != nil {
+		if !canSkipEBPFError(r) {
+			return err
+		}
+	}
+	return nil
 }
